+++ packages/riverpod/test/old/legacy_providers/future_provider/auto_dispose_family_future_provider_test.dart
@@ -2,8 +2,8 @@ import 'package:riverpod/riverpod.dart';
 import 'package:riverpod/src/framework.dart';
 import 'package:test/test.dart';
 
-import '../../core/provider_container_test.dart';
-import '../../../old/utils.dart';
+import '../../../new/core/provider_container_test.dart';
+import '../../utils.dart';
 
 void main() {
   test('specifies `from` & `argument` for related providers', () {
