+++ packages/riverpod/test/old/legacy_providers/stream_provider/auto_dispose_stream_provider_test.dart
@@ -6,7 +6,7 @@ import 'package:riverpod/riverpod.dart';
 import 'package:riverpod/src/internals.dart' show ProviderElementBase;
 import 'package:test/test.dart';
 
-import '../../../old/utils.dart';
+import '../../utils.dart';
 
 void main() {
   group('StreamProvider.autoDispose', () {
