+++ packages/riverpod/test/old/legacy_providers/stream_provider/stream_provider_test.dart
@@ -8,7 +8,7 @@ import 'package:riverpod/riverpod.dart';
 import 'package:riverpod/src/internals.dart' show ProviderElementBase;
 import 'package:test/test.dart';
 
-import '../../../old/utils.dart';
+import '../../utils.dart';
 
 void main() {
   test('supports overrideWith', () {
